separate utils pkg into milestones pkg and args pkg
