+++ close-milestone/close_milestone.go
@@ -3,19 +3,21 @@ package main
 import (
 	"context"
 	"fmt"
-	"grafana-github-actions-go/utils"
 	"log"
 	"os"
+
+	"grafana-github-actions-go/args"
+	"grafana-github-actions-go/milestones"
 
 	gh "github.com/google/go-github/v47/github"
 	"golang.org/x/oauth2"
 )
 
-func updateMilestone(ctx context.Context, editor utils.CloseMilestoneClient, currentVersion string, milestone *gh.Milestone) error {
+func updateMilestone(ctx context.Context, editor milestones.CloseMilestoneClient, currentVersion string, milestone *gh.Milestone) error {
 	// Update milestone status to "closed"
 	milestone.State = gh.String("closed")
 
-	_, _, err := editor.EditMilestone(ctx, "grafana", utils.RepoName, *milestone.Number, milestone)
+	_, _, err := editor.EditMilestone(ctx, "grafana", milestones.RepoName, *milestone.Number, milestone)
 	if err != nil {
 		return fmt.Errorf("did not find milestone: %s", milestone.String())
 	}
@@ -23,7 +25,7 @@ func updateMilestone(ctx context.Context, editor utils.CloseMilestoneClient, cur
 }
 
 func main() {
-	token, currentVersion, err := utils.ReadArgs(os.Args)
+	token, currentVersion, err := args.ReadArgs(os.Args)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -31,7 +33,7 @@ func main() {
 	ctx := context.Background()
 	client := gh.NewClient(oauth2.NewClient(ctx, oauth2.StaticTokenSource(&oauth2.Token{AccessToken: token})))
 
-	milestone, err := utils.FindMilestone(ctx, client.Issues, currentVersion)
+	milestone, err := milestones.FindMilestone(ctx, client.Issues, currentVersion)
 	if err != nil {
 		log.Fatal(err)
 	}
